Calculate ATP Production from OXPHOS and Glycolysis

Calculates ATP production from glycolysis and OXPHOS at points defined in patitioned_data

Usage

get_energetics(partitioned_data, ph, pka, buffer)

Arguments

partitioned_data

a data.table of organized Seahorse OCR and ECAR rates based on timepoints from the assay cycle. Returned by partition_data

ph

pH value for energetics calculation (for XF Media, 7.5)

pka

pKa value for energetics calculation (for XF Media, 6.063)

buffer

buffer for energetics calculation (for XF Media, 0.1 mpH/pmol H+)

Value

a data.table of glycolysis and OXPHOS rates

Details

TODO: check that all symbols are defined

Proton production rate (PPR): $$\text{PPR} = \frac{\text{ECAR value}}{\text{buffer}}$$

$$ \text{PPR}_{\text{mito}} = \frac{10^{\text{pH}-\text{pK}_a}}{1+10^{\text{pH}-\text{pK}_a}} \cdot \frac{\text{H}^+}{\text{O}_2} \cdot \text{OCR} $$

calculates the proton production from glucose during its conversion to bicarbonate and \(\text{H}^+\) assuming max \(\frac{\text{H}^+}{\text{O}_2}\) of 1

$$ \text{PPR}_\text{glyc} = \text{PPR} - \text{PPR}_\text{resp} $$

calculates the proton production from glucose during its conversion to lactate + \(\text{H}^+\)

Joules of ATP (JATP) production:

$$ \text{ATP}_{\text{glyc}} = \Bigl(\text{PPR}_\text{glyc} \cdot \frac{\text{ATP}}{\text{lactate}}\Bigl) + \Bigl(\text{MITO}_\text{resp} \cdot 2 \cdot \frac{\text{P}}{\text{O}_\text{glyc}}\Bigl) $$

$$ \frac{\text{ATP}}{\text{lactate}} = 1 $$ with \(\frac{\text{P}}{{\text{O}_\text{glyc}}}\) = 0.167 for glucose (0.242 for glycogen).

$$ \text{ATP}_\text{resp} = \Bigl(\text{coupled MITO}_\text{resp} \cdot 2 \cdot \frac{\text{P}}{\text{O}_\text{oxphos}}\Bigl) + \Bigl(\text{MITO}_\text{resp} \cdot 2 \cdot \frac{\text{P}}{\text{O}_\text{TCA}}\Bigl) $$ with \(\frac{\text{P}}{{\text{O}_\text{oxphos}}}\) = 2.486 and \(\frac{\text{P}}{{\text{O}_\text{TCA}}}\) = 0.167.

Examples

rep_list <- system.file("extdata", package = "ceas") |>
  list.files(pattern = "*.xlsx", full.names = TRUE)
seahorse_rates <- read_data(rep_list, sheet = 2)
partitioned_data <- partition_data(seahorse_rates)
energetics <- get_energetics(partitioned_data, ph = 7.4, pka = 6.093, buffer = 0.1)
head(energetics, n = 10)
#>     cell_line ATP_basal_resp ATP_max_resp ATP_basal_glyc ATP_max_glyc
#>  1:   Group_1       903.4178     944.6157      176.42439     524.6518
#>  2:   Group_1       900.2175     937.5181      129.69559     442.5421
#>  3:   Group_1      1027.9860    1072.8032      141.43402     472.4431
#>  4:   Group_1       835.1684     875.9062       95.13176     390.3254
#>  5:   Group_1       888.6888     929.8841       87.08441     420.6239
#>  6:   Group_1       907.6305     946.2010      107.80548     485.7402
#>  7:   Group_1       898.6137     934.9272      127.14102     469.7935
#>  8:   Group_1       939.3941     984.5487      114.74865     477.7720
#>  9:   Group_1       852.0317     888.4355      137.66492     504.0603
#> 10:   Group_1       960.6956    1003.8847      103.91464     426.4569