Organize Seahorse Data

Organizes Seahorse OCR and ECAR rates based on defined time points (i.e. the Measurement column) during the experiment. This time point can be specified if you are modifying the Mito and Glyco Stress Test (i.e. from 3 measurements per cycle to X measurements)

partition_data(
  seahorse_rates,
  basal_tp = 3,
  oligo_tp = 6,
  maxresp_tp = 8,
  nonmito_tp = 12,
  maxgly_tp = 6,
  fccp_ecar_tp = 8,
  basal_ecar_tp = 3,
  oligomon_ecar_tp = 8
)

Arguments

seahorse_rates

data.table Seahorse OCR and ECAR rates (imported using read_data function)

basal_tp

Must be less than oligo_tp

oligo_tp

Must be less than maxresp_tp

maxresp_tp

Must be less than nonmito_tp

nonmito_tp

Must be larger than maxresp_tp

maxgly_tp

Must be the same as oligo_tp

fccp_ecar_tp

Must be the same as maxresp_tp

basal_ecar_tp

Must be the same as basal_tp

oligomon_ecar_tp

Must be larger than basal_ecar_tp

Value

a list of named timepoints from each assay cycle

Examples

rep_list <- system.file("extdata", package = "ceas") |> list.files(pattern = "*.xlsx", full.names = TRUE)
seahorse_rates <- read_data(rep_list, sheet = 2)
partitioned_data <- partition_data(seahorse_rates)