Read Seahorse Wave Excel File

Reads input seahore data from an excel Seahorse Wave File

Usage

read_data(rep_list, sheet = 2)

Arguments

rep_list

A list of Seahorse Wave excel export files. One file per replicate. If your data is in a directory called "seahorse_data", use list.files("seahorse_data", pattern = "*.xlsx", full.names = TRUE) to make a list of the excel files.

sheet

The number of the excel sheet containing the long-form Seahorse data. Default is 2 because the long-form output from Seahorse Wave is on sheet 2

Value

a seahorse_rates table

Examples

rep_list <- system.file("extdata", package = "ceas") |>
  list.files(pattern = "*.xlsx", full.names = TRUE)
seahorse_rates <- read_data(rep_list, sheet = 2)
head(seahorse_rates, n = 10)
#>     Measurement Well     Time      OCR     ECAR      PER  exp_group assay_type
#>  1:           1  A01 1.304765   0.0000  0.00000   0.0000 Background       <NA>
#>  2:           1  A02 1.304765 305.2426 30.64529 334.4771    Group_1       MITO
#>  3:           1  A03 1.304765 307.9862 33.27668 358.4754    Group_1       MITO
#>  4:           1  A04 1.304765 339.3399 49.17751 503.4910    Group_2       MITO
#>  5:           1  A05 1.304765 321.9398 47.94602 492.2597    Group_2       MITO
#>  6:           1  A06 1.304765 323.7962 46.84232 482.1940    Group_2       MITO
#>  7:           1  A07 1.304765 379.1455 46.81741 481.9668    Group_3       MITO
#>  8:           1  A08 1.304765 391.1478 50.14648 512.3280    Group_3       MITO
#>  9:           1  A09 1.304765 393.4523 52.54649 534.2160    Group_3       MITO
#> 10:           1  A10 1.304765 217.0543 29.11793 320.5476    Group_4       MITO
#>     replicate
#>  1:         1
#>  2:         1
#>  3:         1
#>  4:         1
#>  5:         1
#>  6:         1
#>  7:         1
#>  8:         1
#>  9:         1
#> 10:         1